moved the triggered/not triggered information to lua

+++ Levels/scripts/deleteRandomObject.lua
@@ -1,14 +1,20 @@
+if(triggeredDeleteRandomObject == nil) then
+    triggeredDeleteRandomObject = false
+end
 function trigger(objectToChange)
-    if(not level) then
-        print("No level found in Lua!")
-        return
+    if(triggeredDeleteRandomObject == false) then
+        if(not level) then
+            print("No level found in Lua!")
+            return
+        end
+        if(not physics) then
+            print("No physics found in Lua!")
+            return
+        end
+        a = level:getObjectCount()
+        rand = math.random(0, a - 1)
+        level:deleteObject(rand)
+        triggeredDeleteRandomObject = true
+        print("deleteRandomObject")
     end
-    if(not physics) then
-        print("No physics found in Lua!")
-        return
-    end
-    a = level:getObjectCount()
-    rand = math.random(0, a - 1)
-    level:deleteObject(rand)
-    print("Triggered from Lua!")
 end
